split graphics
